+++ framework/Database/Migrations/2024_01_10_081049_seed_user_and_roles.php
@@ -1,0 +1,12 @@
+<?php
+
+use Framework\Database\Migration;
+use Framework\Database\Seeders\UserSeeder;
+
+return new class extends Migration
+{
+    public function run(): void
+    {
+        (new UserSeeder())->run();
+    }
+};